[schematest D1R9] Re-prove the complete Delivery 1 contract (#206) (#218)
* schematest: remove structural count shortcuts

* schematest: add compound Build acceptance matrix

* schematest: record final Delivery 1 acceptance

* schematest: fix final acceptance review findings

* schematest: complete cycle 2 acceptance fixes

* schematest: complete cycle 3 review fixes

* schematest: resolve final cycle 3 findings

+++ pkg/schematest/fault_basic.go
@@ -92,7 +92,7 @@ func (machines faultSearchMachines) directParentAtRank(
 
 	groups := parentReplayGroups(fault.requirements)
 
-	decoder, ok := newDirectRankTupleDecoder(len(groups)+1, rank)
+	decoder, ok := newDirectRankTupleDecoder(uint64(len(groups)+1), rank)
 	if !ok {
 		return nil, false, false, nil
 	}
